Moved projectile-related classes to their own namespace
