+++ src/pocketmine/entity/projectile/ProjectileSource.php
@@ -21,7 +21,7 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\entity;
+namespace pocketmine\entity\projectile;
 
 
 interface ProjectileSource{
