+++ src/pocketmine/event/entity/EntitySpawnEvent.php
@@ -27,7 +27,7 @@ use pocketmine\entity\Creature;
 use pocketmine\entity\Entity;
 use pocketmine\entity\Human;
 use pocketmine\entity\Item;
-use pocketmine\entity\Projectile;
+use pocketmine\entity\projectile\Projectile;
 use pocketmine\entity\Vehicle;
 use pocketmine\level\Position;
