+++ src/pocketmine/event/entity/ProjectileLaunchEvent.php
@@ -23,7 +23,7 @@ declare(strict_types=1);
 
 namespace pocketmine\event\entity;
 
-use pocketmine\entity\Projectile;
+use pocketmine\entity\projectile\Projectile;
 use pocketmine\event\Cancellable;
 
 class ProjectileLaunchEvent extends EntityEvent implements Cancellable{
