+++ src/pocketmine/event/inventory/InventoryPickupArrowEvent.php
@@ -23,7 +23,7 @@ declare(strict_types=1);
 
 namespace pocketmine\event\inventory;
 
-use pocketmine\entity\Arrow;
+use pocketmine\entity\projectile\Arrow;
 use pocketmine\event\Cancellable;
 use pocketmine\inventory\Inventory;
